+++ app/seeds/__init__.py
@@ -1,2 +1,3 @@
 from app.seeds.ingredient_seeder import IngredientTemplate
 from app.seeds.beverage_seeder import BeverageTemplate
+from app.seeds.size_seeder import SizeTemplate
